requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
ffmpeg-python==0.2.0
future==0.18.2
greenlet==1.1.2
idna==3.3
imageio-ffmpeg==0.4.7
multidict==6.0.2
numpy==1.23.0
peewee==3.15.1
psycopg2==2.9.3
py-cord==2.0.0
python-dotenv==0.20.0
typing_extensions==4.3.0
yarl==1.7.2
